requirements.txt
# Core dependencies
openai>=1.0.0
elasticsearch>=8.0.0
requests>=2.31.0

# PDF processing
pymupdf>=1.23.0
pdfplumber>=0.10.0
Pillow>=10.0.0

# LangChain
langchain>=0.1.0
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1

# Text processing
tiktoken>=0.5.0
nltk>=3.8.1
beautifulsoup4>=4.12.0
lxml>=4.9.3

# CLI and utilities
python-dotenv>=1.0.0
tqdm>=4.66.0

